% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_embedding.R
\name{sem.emb}
\alias{sem.emb}
\title{Structural Equation Modeling with Embeddings}
\usage{
sem.emb(
  sem_model,
  data,
  text_var,
  encoder = "all-mpnet-base-v2",
  emb_filepath = NULL,
  reduce_method = "SVD",
  reduce_dim = 5
)
}
\arguments{
\item{sem_model}{A character string specifying the SEM model.}

\item{data}{A data frame containing the input data.}

\item{text_var}{A character string specifying the name of the text variable in the data frame.}

\item{encoder}{A character string specifying the encoder model to be used for generating embeddings. Defaults to "all-mpnet-base-v2".}

\item{emb_filepath}{A character string specifying the path to the `.rda` file containing the embeddings. If `NULL`, embeddings are generated using `sem.encode`.}

\item{reduce_dim}{An integer denoting the size of embedding after reduction.}

\item{pca_dim}{A integer specifying reduced dimension through PCA.}
}
\value{
The result of the `lavaan::sem` function, which is an object of class `lavaan`.
}
\description{
This function performs Structural Equation Modeling (SEM) using text embeddings. It checks if the specified `.rda` file with embeddings exists. If the file exists, it loads the embeddings; otherwise, it generates the embeddings using the `sem.encode` function. The embeddings are then incorporated into the SEM model.
}
\examples{
\dontrun{
sem_model <- 'rating ~ book + difficulty + comments'
res <- sem.emb(sem_model = sem_model, data = prof.nest, text_var = "comments",
                pca_dim = 10, emb_model = "all-mpnet-base-v2")
summary(res, fit=TRUE)
}

}
