% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_analysis.R
\name{sem.sentiment}
\alias{sem.sentiment}
\title{Structural Equation Modeling with Sentiment Analysis}
\usage{
sem.sentiment(
  model,
  df,
  text_vars,
  method = "sentimentr",
  text_stats = c("sentiment"),
  polarity_dt = lexicon::hash_sentiment_jockers_rinker,
  valence_shifters_dt = lexicon::hash_valence_shifters,
  missing = "ML",
  fixed.x = FALSE,
  ...
)
}
\arguments{
\item{model}{The structural equation model specified as a character string.}

\item{df}{A data frame containing the input data.}

\item{text_vars}{A character vector of text variable names in the data frame for which sentiment analysis should be performed.}

\item{text_stats}{A character vector of text sentiment statistics to be added to the SEM. Currently supports only 'OverallSenti' (overall sentiment).}

\item{polarity_dt}{A data table for polarity lexicon to be used for sentiment analysis. Defaults to `lexicon::hash_sentiment_jockers_rinker`.}

\item{valence_shifters_dt}{A data table for valence shifters to be used for sentiment analysis. Defaults to `lexicon::hash_valence_shifters`.}

\item{missing}{The method for handling missing data in the SEM. Defaults to 'ML' (maximum likelihood).}

\item{fixed.x}{Logical. If `TRUE`, the exogenous variables are treated as fixed. Defaults to `FALSE`.}

\item{...}{Additional arguments passed to the `sentiment_by` function from the `sentimentr` package and to the `sem` function from the `lavaan` package.}
}
\value{
A list containing three items:
\item{model}{A character string representing the modified SEM with added sentiment variables.}
\item{data}{A data frame with added text sentiment statistics.}
\item{estimates}{The fitted SEM model object.}
}
\description{
This function integrates sentiment analysis into a structural equation model (SEM) by calculating sentiment scores for specified text variables and incorporating them as additional variables in the SEM.
}
