% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{sem.topic.path}
\alias{sem.topic.path}
\title{Plot SEM Path Diagram with topic model}
\usage{
sem.topic.path(sem.model)
}
\arguments{
\item{sem.model}{A fitted SEM model object.}
}
\value{
A rendered graph visualization of the SEM path diagram.
}
\description{
This function plots the path diagram of a Structural Equation Modeling (SEM) with topic model.
}
\examples{
\dontrun{
# Assuming 'fit' is a fitted SEM model object
sem.topic.path(fit)
}
}
