% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{sem.topic.plot}
\alias{sem.topic.plot}
\title{Plot Top Terms in LDA Topics}
\usage{
sem.topic.plot(df.lda)
}
\arguments{
\item{df.lda}{A fitted LDA model object.}
}
\value{
A ggplot object showing the top terms in each topic.
}
\description{
This function plots the top terms in each topic from a Latent Dirichlet Allocation (LDA) model.
}
\examples{
\dontrun{
# Assuming 'lda_model' is a fitted LDA model object
sem.topic.plot(lda_model)
}
}
