% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{sem.topic}
\alias{sem.topic}
\title{Perform Structural Equation Modeling with Latent Dirichlet Allocation}
\usage{
sem.topic(df, id_var, text_var, n_topics, model)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{id_var}{A variable in the data frame that uniquely identifies each document.}

\item{text_var}{A variable in the data frame containing the text data to be analyzed.}

\item{n_topics}{Number of topics to be extracted.}

\item{model}{A description of the user-specified model. Typically, the model is described using the lavaan model syntax. See model.syntax for more information. Alternatively, a parameter table (eg. the output of the lavaanify() function) is also accepted.}
}
\value{
A list containing two elements: `model` (the fitted SEM model) and `lda` (the LDA model).
}
\description{
This function performs structural equation modeling (SEM) combined with Latent Dirichlet Allocation (LDA) to analyze text data.
}
\examples{
\dontrun{
data(prof.nest)
res <- sem.topic(prof.nest, 'profid', 'comments')
summary(res$model, fit=TRUE)
model <- 'rating ~ book + difficulty +
          topic1 + topic2 + topic3 + topic4 + topic5
         '
}
}
