% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSEM-package.R
\docType{package}
\name{TextSEM-package}
\alias{TextSEM}
\alias{TextSEM-package}
\title{TextSEM: A framework for combining text data in structural equation modeling (SEM)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Stan7s/TextSEM}
  \item Report bugs at \url{https://github.com/Stan7s/TextSEM/issues}
}

}
\keyword{internal}
