% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{textsem_init}
\alias{textsem_init}
\title{Initialize Environment}
\usage{
textsem_init(envname = "r-reticulate")
}
\arguments{
\item{envname}{A character string specifying the name of the virtual environment previously created. Default is `"r-reticulate"`.}
}
\value{
This function is used for side effects: it activates the virtual environment and initializes the `sentiment.ai` library.
}
\description{
The `textsem_init` function activates the virtual environment created by `textsem_install` and initializes the `sentiment.ai` library, enabling sentiment analysis capabilities in R.
}
\details{
This function activates the specified virtual environment and initializes necessary libraries for conducting text sentiment analysis. It is typically called after `textsem_install` to prepare the environment for specific analysis sessions.
}
\examples{
\dontrun{
# Activate and initialize the environment
textsem_init()
}

}
