% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{textsem_install}
\alias{textsem_install}
\title{Install Environment}
\usage{
textsem_install(envname = "r-reticulate")
}
\arguments{
\item{envname}{A character string specifying the name of the virtual environment. Default is `"r-reticulate"`.}
}
\value{
This function is used for side effects: it creates a virtual environment and installs necessary Python packages.
}
\description{
This function, `textsem_install`, is designed to create a Python virtual environment and install essential Python packages required for advanced text semantics analysis.
}
\details{
The function sets up a virtual environment and installs Python packages commonly used in machine learning and natural language processing. These include `pandas`, `torch`, `transformers`, and several other libraries tailored for specific tasks in text analysis and machine learning.
}
\examples{
\dontrun{
# Install the environment and necessary packages
textsem_install()
}

}
