% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_embedding.R
\name{sem.emb}
\alias{sem.emb}
\title{Structural Equation Modeling with Embeddings}
\usage{
sem.emb(
  df,
  text_var,
  sem_model,
  emb_file_path = NULL,
  encoder_model = "all-mpnet-base-v2"
)
}
\arguments{
\item{df}{A data frame containing the input data.}

\item{text_var}{A character string specifying the name of the text variable in the data frame.}

\item{sem_model}{A character string specifying the SEM model.}

\item{emb_file_path}{A character string specifying the path to the `.rda` file containing the embeddings. If `NULL`, embeddings are generated using `sem.encode`.}

\item{encoder_model}{A character string specifying the encoder model to be used for generating embeddings. Defaults to "all-mpnet-base-v2".}
}
\value{
The result of the `lavaan::sem` function, which is an object of class `lavaan`.
}
\description{
This function performs Structural Equation Modeling (SEM) using text embeddings. It checks if the specified `.rda` file with embeddings exists. If the file exists, it loads the embeddings; otherwise, it generates the embeddings using the `sem.encode` function. The embeddings are then incorporated into the SEM model.
}
\examples{
\dontrun{
df <- data.frame(text = c("I love this!", "I hate that!"), rating = c(5, 1))
sem_model <- "rating ~ text"
result <- sem.emb(df, text_var = "text", sem_model = sem_model, emb_file_path = "embeddings.rda")
summary(result)
}
}
