% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_embedding.R
\name{sem.encode}
\alias{sem.encode}
\title{Generate Sentence Embeddings using Sentence Transformers}
\usage{
sem.encode(text_vector, model_name = "all-mpnet-base-v2")
}
\arguments{
\item{text_vector}{A character vector containing the text data to be embedded.}

\item{model_name}{A character string specifying the name of the pre-trained model to be used for generating embeddings.}
}
\value{
A matrix of sentence embeddings with each row corresponding to a sentence from `text_vector` and each column representing a dimension of the embedding space.
}
\description{
This function generates sentence embeddings for a given vector of text using a specified pre-trained model from the `sentence_transformers` Python package.
}
\examples{
\dontrun{
# Example usage
text_vector <- c("This is a sentence.", "This is another sentence.")
model_name <- "paraphrase-MiniLM-L6-v2"
embeddings <- sem.emb(text_vector, model_name)
print(embeddings)
}
}
