% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{sem.lda}
\alias{sem.lda}
\title{Perform Latent Dirichlet Allocation on a Data Frame}
\usage{
sem.lda(df, id_var, text_var)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{id_var}{A variable in the data frame that uniquely identifies each document.}

\item{text_var}{A variable in the data frame containing the text data to be analyzed.}
}
\value{
A topic model object of class "LDA" from the `topicmodels` package.
}
\description{
This function takes a data frame and performs text preprocessing followed by Latent Dirichlet Allocation (LDA) for topic modeling.
}
\examples{
\dontrun{
data(prof.nest)
lda.model <- sem.lda(df, id_var = "profid", text_var = "comments")
lda.model
}
}
