% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{sem.lda}
\alias{sem.lda}
\title{Perform Latent Dirichlet Allocation on a Data Frame}
\usage{
sem.lda(data, text_var, n_topic, method = "VEM", sparse = 0.995, seed = 42)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{text_var}{A variable in the data frame containing the text data to be analyzed.}

\item{n_topic}{Number of topics to be extracted.}

\item{method}{The method to be used for LDA fitting; currently method = "VEM" or method= "Gibbs" are supported.}

\item{sparse}{A numeric for the maximal allowed sparsity in the range from bigger zero to smaller one.}

\item{seed}{Random seed for LDA estimation}
}
\value{
A topic model object of class "LDA" from the `topicmodels` package.
}
\description{
This function takes a data frame and performs text preprocessing followed by Latent Dirichlet Allocation (LDA) for topic modeling.
}
\examples{
\dontrun{
data(prof.nest)
lda.model <- sem.lda(df, text_var = c("comments"), n_topic = c(6))
lda.model
}
}
