% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_model.R
\name{sem.topic}
\alias{sem.topic}
\title{Perform Structural Equation Modeling with Latent Dirichlet Allocation}
\usage{
sem.topic(
  model,
  data,
  text_vars,
  n_topics,
  method = "VEM",
  sparse = 0.995,
  seed = 42
)
}
\arguments{
\item{model}{A description of the user-specified model. Typically, the model is described using the lavaan model syntax. See model.syntax for more information. Alternatively, a parameter table (eg. the output of the lavaanify() function) is also accepted.}

\item{data}{A data frame containing the data.}

\item{text_vars}{A character vector of text variable names in the data frame containing the text data to be analyzed.}

\item{n_topics}{A numeric vector containing number of topics to be extracted for each text variable.}

\item{method}{The method to be used for LDA fitting; currently method = "VEM" or method= "Gibbs" are supported.}

\item{sparse}{A numeric for the maximal allowed sparsity in the range from bigger zero to smaller one.}

\item{seed}{Random seed for LDA estimation}
}
\value{
A list containing four elements:
\item{model}{A character string representing the modified SEM with added topic variables.}
\item{data}{A data frame with added topic statistics.}
\item{estimates}{The fitted SEM model object.}
\item{lda}{A vector of LDA model objects.}
}
\description{
This function performs structural equation modeling (SEM) combined with Latent Dirichlet Allocation (LDA) to analyze text data.
}
\examples{
\dontrun{
data(prof.nest)
model <- 'rating ~ book + difficulty + comments + tags'
res <- sem.topic(model = model,
                 data = prof.nest,
                 text_vars = c('comments', 'tags'),
                 n_topics = c(6, 3))
summary(res$model, fit=TRUE)
}
}
